EDU-1412

* EDU-1412:Changed condition

* EDU-1442:propert changes

Approved-by: Karan Verma <karan2.verma@paytmmall.com>

+++ digital-education-explore/src/main/java/com/paytm/digital/education/explore/service/impl/CompareServiceImpl.java
@@ -397,8 +397,8 @@ public class CompareServiceImpl implements CompareService {
 
     private List<Course> getCourses(Long instId) {
         List<Course> courses = new ArrayList<>();
-        if (Objects.nonNull(instituteCoursesMap.get(instId.intValue()))) {
-            courses = instituteCoursesMap.get(instId.intValue());
+        if (Objects.nonNull(instituteCoursesMap.get(instId))) {
+            courses = instituteCoursesMap.get(instId);
         }
         return courses;
     }
